requirements.txt
PlexAPI==4.13.4
requests==2.28.2
python-dotenv==1.0.0
aiohttp==3.8.4
loguru==0.7.0
